add in simplecov and refactor forme configs to formi namespace and reorg for eager loading
