+++ .gitignore
@@ -2,6 +2,6 @@
 /repos
 /logs
 /node_modules
-
+/coverage
 
 test.fossil
