+++ lib/formi/bones/wrapper.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Formi::Bones::Wrapper < Forme::Wrapper
+  Forme.register_transformer :wrapper, :bones, new
+
+  def call tag, input
+    case input.type
+    when :submit, :button, :reset then [tag]
+    else
+      [ input.tag(:div, { class: "m-form__input-group" }, tag) ]
+    end
+  end
+end
